refactor skymodel module. first steps.
